# contracts/openapi/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ChatRequest:
    properties:
      context:
        additionalProperties: true
        type: object
      message:
        example: Quanto de água eu bebi hoje?
        maxLength: 2000
        minLength: 1
        type: string
      runtime:
        $ref: '#/definitions/dto.ChatRuntimeSelection'
    required:
    - message
    type: object
  dto.ChatResponse:
    properties:
      response:
        example: Você bebeu 2.5 litros de água hoje...
        type: string
      sources:
        items:
          additionalProperties: true
          type: object
        type: array
      ui:
        additionalProperties: true
        type: object
      usage:
        $ref: '#/definitions/dto.TokenUsage'
    type: object
  dto.ChatRuntimeSelection:
    properties:
      model:
        example: gpt-5.4-mini
        type: string
      provider:
        example: openai
        type: string
    type: object
...
  dto.TokenUsage:
    properties:
      completion_tokens:
        example: 100
        type: integer
      prompt_tokens:
        example: 50
        type: integer
      total_tokens:
        example: 150
        type: integer
    type: object
...
info:
  contact:
    email: felipe.lechitz@gmail.com
    name: Lechitz
    url: https://github.com/lechitz
  description: Public and internal REST API for aion-api. Swagger (OpenAPI 3.x) is
    generated via swaggo.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/lechitz/aion-api
  title: aion-api — REST API Documentation
  version: 0.0.1
paths:
...
  /chat/text:
    post:
      consumes:
      - application/json
      description: Sends a message to the AI assistant and receives a response. Requires
        authentication.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ChatText message
        in: body
        name: chat
        required: true
        schema:
          $ref: '#/definitions/dto.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ChatText response
          schema:
            $ref: '#/definitions/dto.ChatResponse'
        "400":
          description: Invalid request payload or validation error
          schema:
            type: string
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
        "503":
          description: Service unavailable - AI service is down
          schema:
            type: string
      security:
      - BearerAuth.: []
      summary: Send chat message
      tags:
      - ChatText
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Use "Bearer {token}" (JWT).
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: Session-based authentication cookie (e.g., Redis-backed).
    in: cookie
    name: auth_token
    type: apiKey
    x-graphqlplayground: /graphql/ // Cross-reference to GraphQL Playground (not covered
      by Swagger)
swagger: "2.0"